add placement specs to page spac in 2 different describe
